Case:32794 More refactoring as Layout interaction was broken.  FormatBuilder now supports ISynchronizedTextWriter and IControllableTextWriter, whilst LayoutBuilder understands ILayoutTextWriter.

+++ Utilities/WebApplications.Utilities.Test/Formatting/TestLayoutBuilder.cs
@@ -70,7 +70,7 @@ namespace WebApplications.Utilities.Test.Formatting
                     tabStops: new ushort[] {6, 9, 20, 30, 40})
                 .Append("A\tTab Stop\tAnother");
 
-            int position = 0;
+            ushort position = 0;
             Assert.AreEqual(" A    Tab Stop      Another", builder.ToString(ref position));
             Assert.AreEqual(27, position);
             Assert.AreEqual("A  Tab Stop  Another", builder.ToString(ref position));
